+++ RevenueCatUI/Templates/Components/ComponentViewModels.swift
@@ -145,148 +145,6 @@ extension PaywallComponent.Padding {
         }
 }
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-public class TextComponentViewModel: ObservableObject {
-    let locale: Locale
-    let localization: [String: String]
-    @Published private(set) var component: PaywallComponent.TextComponent
-
-    init(locale: Locale, localization: [String : String], component: PaywallComponent.TextComponent) {
-        self.locale = locale
-        self.localization = localization
-        self.component = component
-    }
-
-    var text: String {
-        // TODO: Replace variables like "{{ }}"
-        // TODO: Add logs?
-        if let textLid = component.textLid {
-            if let localizedText = localization[textLid] {
-                return localizedText
-            }
-            else {
-                return component.text.value.first?.value as? String ?? "missing localized text for \(textLid)"
-            }
-        } else {
-            return component.text.value.first?.value as? String ?? "missing localized text"
-        }
-    }
-
-    public var fontFamily: String {
-        component.fontFamily
-    }
-
-    public var fontWeight: Font.Weight {
-        component.fontWeight.fontWeight
-    }
-
-    public var color: Color {
-        // TODO: implement color transformation
-        // component.color
-        Color.cyan
-    }
-
-    public var textStyle: Font {
-        component.textStyle.font
-    }
-
-    public var horizontalAlignment: TextAlignment {
-        component.horizontalAlignment.textAlignment
-    }
-
-    public var backgroundColor: Color {
-        // TODO: implement color transformation
-        // component.color
-        Color.mint
-    }
-
-    public var padding: EdgeInsets {
-        component.padding.edgeInsets
-    }
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-// @PublicForExternalTesting
-class ImageComponentViewModel: ObservableObject {
-    let locale: Locale
-    @Published private(set) var component: PaywallComponent.ImageComponent
-
-    init(locale: Locale, component: PaywallComponent.ImageComponent) {
-        self.locale = locale
-        self.component = component
-    }
-    
-    public var url: URL {
-        component.url
-    }
-    public var cornerRadius: Double {
-        component.cornerRadius
-    }
-    public var gradientColors: [Color] {
-        component.gradientColors.compactMap { try? $0.toColor() }
-    }
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-// @PublicForExternalTesting
-class SpacerComponentViewModel: ObservableObject {
-    let locale: Locale
-    let component: PaywallComponent.SpacerComponent
-
-    init(locale: Locale, component: PaywallComponent.SpacerComponent) {
-        self.locale = locale
-        self.component = component
-    }
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-// @PublicForExternalTesting
-class StackComponentViewModel: ObservableObject {
-    let locale: Locale
-    let component: PaywallComponent.StackComponent
-    let viewModels: [PaywallComponentViewModel]
-
-    init(locale: Locale,
-         component: PaywallComponent.StackComponent,
-         localization: [String: String],
-         offering: Offering
-    ) {
-        self.locale = locale
-        self.component = component
-        self.viewModels = component.components.map {
-            $0.toViewModel(offering: offering, locale: locale, localization: localization)
-        }
-
-    }
-
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-// @PublicForExternalTesting
-class LinkButtonComponentViewModel: ObservableObject {
-    let locale: Locale
-    @Published private(set) var component: PaywallComponent.LinkButtonComponent
-    let textComponentViewModel: TextComponentViewModel
-
-    public var url: URL {
-        component.url
-    }
-    public var textComponent: PaywallComponent.TextComponent {
-        component.textComponent
-    }
-
-    init(locale: Locale,
-         component: PaywallComponent.LinkButtonComponent,
-         localization: [String: String],
-         offering: Offering
-    ) {
-        self.locale = locale
-        self.component = component
-        self.textComponentViewModel = TextComponentViewModel(locale: locale, localization: localization, component: component.textComponent)
-
-    }
-
-}
 
 @available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
 // @PublicForExternalTesting
